+++ packages/flutter/lib/parse_server_sdk_flutter.dart
@@ -23,8 +23,8 @@ export 'package:parse_server_sdk/parse_server_sdk.dart'
     hide Parse, CoreStoreSembastImp;
     
 // Analytics integration
-export 'src/analytics/parse_analytics_clean.dart';
-export 'src/analytics/parse_analytics_endpoints_clean.dart';
+export 'src/analytics/parse_analytics.dart';
+export 'src/analytics/parse_analytics_endpoints.dart';
 
 part 'src/storage/core_store_shared_preferences.dart';
