+++ packages/flutter/lib/src/analytics/parse_analytics_endpoints.dart
@@ -1,349 +1,277 @@
-import 'dart:convert';
 import 'package:flutter/foundation.dart';
-import 'parse_analytics_clean.dart';
+import 'parse_analytics.dart';
 
-/// HTTP endpoints handler for Parse Dashboard Analytics integration
+/// HTTP endpoint handlers for Parse Dashboard Analytics integration
 class ParseAnalyticsEndpoints {
-  static ParseAnalyticsEndpoints? _instance;
-  static ParseAnalyticsEndpoints get instance => _instance ??= ParseAnalyticsEndpoints._();
   
-  ParseAnalyticsEndpoints._();
-  
-  final ParseAnalytics _analytics = ParseAnalytics.instance;
-  
-  /// Handle analytics content audience endpoint
-  /// GET /apps/{appSlug}/analytics_content_audience?at={timestamp}&audienceType={type}
-  Future<Map<String, dynamic>> handleAudienceRequest({
-    required String audienceType,
-    int? timestamp,
-  }) async {
+  /// Handle audience analytics requests for Parse Dashboard
+  static Future<Map<String, dynamic>> handleAudienceRequest(String audienceType) async {
     try {
-      final DateTime? date = timestamp != null 
-          ? DateTime.fromMillisecondsSinceEpoch(timestamp * 1000)
-          : null;
-      
-      // Check if we have valid cached data
-      final cacheKey = 'audience_$audienceType';
-      if (_analytics.isCacheValid(cacheKey)) {
-        final cached = _analytics.getCachedMetrics(cacheKey);
-        if (cached.isNotEmpty) {
-          return {'total': cached['value'] ?? 0, 'content': cached['value'] ?? 0};
-        }
-      }
-      
-      int value = 0;
+      final userAnalytics = await ParseAnalytics.getUserAnalytics();
+      final installationAnalytics = await ParseAnalytics.getInstallationAnalytics();
       
       switch (audienceType) {
         case 'total_users':
-          final userAnalytics = await _analytics.getUserAnalytics();
-          value = userAnalytics['total_users'] ?? 0;
-          break;
+          return {
+            'total': userAnalytics['total_users'],
+            'content': userAnalytics['total_users']
+          };
           
         case 'daily_users':
-          final userAnalytics = await _analytics.getUserAnalytics();
-          value = userAnalytics['daily_users'] ?? 0;
-          break;
+          return {
+            'total': userAnalytics['daily_users'],
+            'content': userAnalytics['daily_users']
+          };
           
         case 'weekly_users':
-          final userAnalytics = await _analytics.getUserAnalytics();
-          value = userAnalytics['weekly_users'] ?? 0;
-          break;
+          return {
+            'total': userAnalytics['weekly_users'],
+            'content': userAnalytics['weekly_users']
+          };
           
         case 'monthly_users':
-          final userAnalytics = await _analytics.getUserAnalytics();
-          value = userAnalytics['monthly_users'] ?? 0;
-          break;
+          return {
+            'total': userAnalytics['monthly_users'],
+            'content': userAnalytics['monthly_users']
+          };
           
         case 'total_installations':
-          final installationAnalytics = await _analytics.getInstallationAnalytics();
-          value = installationAnalytics['total_installations'] ?? 0;
-          break;
+          return {
+            'total': installationAnalytics['total_installations'],
+            'content': installationAnalytics['total_installations']
+          };
           
         case 'daily_installations':
-          final installationAnalytics = await _analytics.getInstallationAnalytics();
-          value = installationAnalytics['daily_installations'] ?? 0;
-          break;
+          return {
+            'total': installationAnalytics['daily_installations'],
+            'content': installationAnalytics['daily_installations']
+          };
           
         case 'weekly_installations':
-          final installationAnalytics = await _analytics.getInstallationAnalytics();
-          value = installationAnalytics['weekly_installations'] ?? 0;
-          break;
+          return {
+            'total': installationAnalytics['weekly_installations'],
+            'content': installationAnalytics['weekly_installations']
+          };
           
         case 'monthly_installations':
-          final installationAnalytics = await _analytics.getInstallationAnalytics();
-          value = installationAnalytics['monthly_installations'] ?? 0;
-          break;
+          return {
+            'total': installationAnalytics['monthly_installations'],
+            'content': installationAnalytics['monthly_installations']
+          };
           
         default:
-          value = 0;
+          return {'total': 0, 'content': 0};
       }
-      
-      // Cache the result
-      _analytics._cachedMetrics[cacheKey] = {
-        'value': value,
-      };
-      _analytics._cachedMetrics['${cacheKey}_timestamp'] = DateTime.now().millisecondsSinceEpoch;
-      
-      return {
-        'total': value,
-        'content': value,
-      };
-      
-    } catch (error) {
+    } catch (e) {
       if (kDebugMode) {
-        print('ParseAnalyticsEndpoints handleAudienceRequest error: $error');
+        print('Error handling audience request: $e');
       }
       return {'total': 0, 'content': 0};
     }
   }
-  
-  /// Handle billing storage endpoint
-  /// GET /apps/{appSlug}/billing_file_storage
-  Future<Map<String, dynamic>> handleFileStorageRequest() async {
-    try {
-      // This is a placeholder - implement based on your storage system
-      // You would typically query your file storage system (AWS S3, etc.)
-      return {
-        'total': 0.0, // Size in GB
-        'limit': 100.0, // Your storage limit in GB
-        'units': 'GB'
-      };
-    } catch (error) {
-      return {
-        'total': 0.0,
-        'limit': 100.0,
-        'units': 'GB'
-      };
-    }
-  }
-  
-  /// Handle database storage endpoint
-  /// GET /apps/{appSlug}/billing_database_storage
-  Future<Map<String, dynamic>> handleDatabaseStorageRequest() async {
-    try {
-      // This is a placeholder - implement based on your database
-      // For MongoDB you might use db.stats(), for PostgreSQL pg_database_size()
-      return {
-        'total': 0.0, // Size in GB
-        'limit': 20.0, // Your database limit in GB
-        'units': 'GB'
-      };
-    } catch (error) {
-      return {
-        'total': 0.0,
-        'limit': 20.0,
-        'units': 'GB'
-      };
-    }
-  }
-  
-  /// Handle data transfer endpoint
-  /// GET /apps/{appSlug}/billing_data_transfer
-  Future<Map<String, dynamic>> handleDataTransferRequest() async {
-    try {
-      // This would need to be tracked by your middleware/proxy
-      return {
-        'total': 0.0, // Size in TB
-        'limit': 1.0, // Your transfer limit in TB
-        'units': 'TB'
-      };
-    } catch (error) {
-      return {
-        'total': 0.0,
-        'limit': 1.0,
-        'units': 'TB'
-      };
-    }
-  }
-  
-  /// Handle analytics time series endpoint
-  /// GET /apps/{appSlug}/analytics?endpoint={type}&audienceType={type}&stride={stride}&from={timestamp}&to={timestamp}
-  Future<Map<String, dynamic>> handleAnalyticsRequest({
+
+  /// Handle analytics time series requests for Parse Dashboard
+  static Future<Map<String, dynamic>> handleAnalyticsRequest({
     required String endpoint,
-    String? audienceType,
-    String stride = 'day',
-    int? from,
-    int? to,
+    required DateTime startDate,
+    required DateTime endDate,
+    String interval = 'day',
   }) async {
     try {
-      final startDate = from != null 
-          ? DateTime.fromMillisecondsSinceEpoch(from * 1000)
-          : DateTime.now().subtract(const Duration(days: 7));
-      
-      final endDate = to != null 
-          ? DateTime.fromMillisecondsSinceEpoch(to * 1000)
-          : DateTime.now();
-      
-      String metricType;
+      String metric;
       switch (endpoint) {
         case 'audience':
-          metricType = 'active_users';
+          metric = 'users';
           break;
         case 'installations':
-          metricType = 'installations';
-          break;
-        case 'api_request':
-          metricType = 'custom_events'; // You can track API requests as custom events
-          break;
-        case 'push':
-          metricType = 'custom_events'; // Track push notifications as custom events
+          metric = 'installations';
           break;
         default:
-          metricType = 'custom_events';
+          metric = endpoint;
       }
       
-      final requestedData = await _analytics.getTimeSeriesData(
-        metricType: metricType,
+      final requestedData = await ParseAnalytics.getTimeSeriesData(
+        metric: metric,
         startDate: startDate,
         endDate: endDate,
-        stride: stride,
+        interval: interval,
       );
       
-      return {
-        'requested_data': requestedData,
-      };
-      
-    } catch (error) {
+      return {'requested_data': requestedData};
+    } catch (e) {
       if (kDebugMode) {
-        print('ParseAnalyticsEndpoints handleAnalyticsRequest error: $error');
+        print('Error handling analytics request: $e');
       }
-      return {
-        'requested_data': <List<dynamic>>[],
-      };
+      return {'requested_data': <List<num>>[]};
     }
   }
-  
-  /// Handle analytics retention endpoint
-  /// GET /apps/{appSlug}/analytics_retention?at={timestamp}
-  Future<Map<String, dynamic>> handleRetentionRequest({
-    int? timestamp,
-  }) async {
+
+  /// Handle user retention requests for Parse Dashboard
+  static Future<Map<String, double>> handleRetentionRequest({DateTime? cohortDate}) async {
     try {
-      final cohortDate = timestamp != null
-          ? DateTime.fromMillisecondsSinceEpoch(timestamp * 1000)
-          : DateTime.now().subtract(const Duration(days: 30));
-      
-      final retention = await _analytics.getUserRetention(cohortDate: cohortDate);
-      
-      return retention;
-      
-    } catch (error) {
+      return await ParseAnalytics.getUserRetention(cohortDate: cohortDate);
+    } catch (e) {
       if (kDebugMode) {
-        print('ParseAnalyticsEndpoints handleRetentionRequest error: $error');
+        print('Error handling retention request: $e');
       }
-      return {
-        'day1': 0.0,
-        'day7': 0.0,
-        'day30': 0.0,
-      };
+      return {'day1': 0.0, 'day7': 0.0, 'day30': 0.0};
     }
   }
-  
-  /// Handle slow queries endpoint
-  /// GET /apps/{appSlug}/slow_queries?className={className}&os={os}&version={version}&from={timestamp}&to={timestamp}
-  Future<Map<String, dynamic>> handleSlowQueriesRequest({
+
+  /// Handle billing storage requests for Parse Dashboard
+  static Map<String, dynamic> handleBillingStorageRequest() {
+    // Mock implementation - replace with actual storage calculation
+    return {
+      'total': 0.5, // 500MB in GB
+      'limit': 100,
+      'units': 'GB'
+    };
+  }
+
+  /// Handle billing database requests for Parse Dashboard  
+  static Map<String, dynamic> handleBillingDatabaseRequest() {
+    // Mock implementation - replace with actual database size calculation
+    return {
+      'total': 0.1, // 100MB in GB
+      'limit': 20,
+      'units': 'GB'
+    };
+  }
+
+  /// Handle billing data transfer requests for Parse Dashboard
+  static Map<String, dynamic> handleBillingDataTransferRequest() {
+    // Mock implementation - replace with actual data transfer calculation
+    return {
+      'total': 0.001, // 1GB in TB
+      'limit': 1,
+      'units': 'TB'
+    };
+  }
+
+  /// Handle slow queries requests for Parse Dashboard
+  static List<Map<String, dynamic>> handleSlowQueriesRequest({
     String? className,
     String? os,
     String? version,
-    int? from,
-    int? to,
-  }) async {
-    try {
-      // This would need to be implemented based on your Parse Server logs
-      // You would analyze query performance logs and return slow queries
-      
-      return {
-        'result': <Map<String, dynamic>>[
-          // Example slow query entry:
-          // {
-          //   'className': '_User',
-          //   'query': '{"username": "example"}',
-          //   'duration': 1500, // milliseconds
-          //   'count': 10, // number of times this query ran slowly
-          // }
-        ],
-      };
-      
-    } catch (error) {
-      return {
-        'result': <Map<String, dynamic>>[],
-      };
-    }
+    DateTime? from,
+    DateTime? to,
+  }) {
+    // Mock implementation - replace with actual slow query analysis
+    return [
+      {
+        'className': className ?? '_User',
+        'query': '{"username": {"regex": ".*"}}',
+        'duration': 1200,
+        'count': 5,
+        'timestamp': DateTime.now().toIso8601String()
+      }
+    ];
+  }
+}
+
+/// Express.js middleware generator for Parse Dashboard integration
+/// 
+/// Usage:
+/// ```javascript
+/// const express = require('express');
+/// const app = express();
+/// app.use('/apps/:appSlug/', getExpressMiddleware());
+/// ```
+String getExpressMiddleware() {
+  return '''
+// Parse Dashboard Analytics middleware
+function parseAnalyticsMiddleware(req, res, next) {
+  // Add authentication middleware here
+  const masterKey = req.headers['x-parse-master-key'];
+  if (!masterKey || masterKey !== process.env.PARSE_MASTER_KEY) {
+    return res.status(401).json({ error: 'Unauthorized' });
   }
   
-  /// Create Express.js/Dart Shelf middleware handlers
-  /// This provides example implementations for common server frameworks
-  Map<String, Function> getExpressMiddleware() {
-    return {
-      'analytics_content_audience': (dynamic req, dynamic res) async {
-        final audienceType = req.query['audienceType'] as String? ?? '';
-        final timestampStr = req.query['at'] as String?;
-        final timestamp = timestampStr != null ? int.tryParse(timestampStr) : null;
-        
-        final result = await handleAudienceRequest(
-          audienceType: audienceType,
-          timestamp: timestamp,
-        );
-        
-        return result;
-      },
-      
-      'billing_file_storage': (dynamic req, dynamic res) async {
-        return await handleFileStorageRequest();
-      },
-      
-      'billing_database_storage': (dynamic req, dynamic res) async {
-        return await handleDatabaseStorageRequest();
-      },
-      
-      'billing_data_transfer': (dynamic req, dynamic res) async {
-        return await handleDataTransferRequest();
-      },
-      
-      'analytics': (dynamic req, dynamic res) async {
-        final endpoint = req.query['endpoint'] as String? ?? '';
-        final audienceType = req.query['audienceType'] as String?;
-        final stride = req.query['stride'] as String? ?? 'day';
-        final fromStr = req.query['from'] as String?;
-        final toStr = req.query['to'] as String?;
-        
-        final from = fromStr != null ? int.tryParse(fromStr) : null;
-        final to = toStr != null ? int.tryParse(toStr) : null;
-        
-        return await handleAnalyticsRequest(
-          endpoint: endpoint,
-          audienceType: audienceType,
-          stride: stride,
-          from: from,
-          to: to,
-        );
-      },
-      
-      'analytics_retention': (dynamic req, dynamic res) async {
-        final timestampStr = req.query['at'] as String?;
-        final timestamp = timestampStr != null ? int.tryParse(timestampStr) : null;
-        
-        return await handleRetentionRequest(timestamp: timestamp);
-      },
-      
-      'slow_queries': (dynamic req, dynamic res) async {
-        final className = req.query['className'] as String?;
-        final os = req.query['os'] as String?;
-        final version = req.query['version'] as String?;
-        final fromStr = req.query['from'] as String?;
-        final toStr = req.query['to'] as String?;
-        
-        final from = fromStr != null ? int.tryParse(fromStr) : null;
-        final to = toStr != null ? int.tryParse(toStr) : null;
-        
-        return await handleSlowQueriesRequest(
-          className: className,
-          os: os,
-          version: version,
-          from: from,
-          to: to,
-        );
-      },
-    };
+  // Route analytics requests
+  if (req.path.includes('analytics_content_audience')) {
+    // Handle audience requests
+    const audienceType = req.query.audienceType;
+    // Call your Flutter analytics endpoint here
+    return res.json({ total: 0, content: 0 });
   }
+  
+  if (req.path.includes('analytics')) {
+    // Handle analytics time series requests
+    const { endpoint, stride, from, to } = req.query;
+    // Call your Flutter analytics endpoint here
+    return res.json({ requested_data: [] });
+  }
+  
+  if (req.path.includes('analytics_retention')) {
+    // Handle retention requests
+    // Call your Flutter analytics endpoint here
+    return res.json({ day1: 0, day7: 0, day30: 0 });
+  }
+  
+  next();
+}
+
+module.exports = parseAnalyticsMiddleware;
+''';
+}
+
+/// Dart Shelf handler for Parse Dashboard integration
+/// 
+/// Usage:
+/// ```dart
+/// import 'package:shelf/shelf.dart';
+/// import 'package:shelf/shelf_io.dart' as io;
+/// 
+/// void main() async {
+///   final handler = getDartShelfHandler();
+///   final server = await io.serve(handler, 'localhost', 3000);
+/// }
+/// ```
+String getDartShelfHandler() {
+  return '''
+import 'dart:convert';
+import 'package:shelf/shelf.dart';
+
+Response Function(Request) getDartShelfHandler() {
+  return (Request request) async {
+    // Add authentication here
+    final masterKey = request.headers['x-parse-master-key'];
+    if (masterKey != 'your_master_key') {
+      return Response.forbidden(json.encode({'error': 'Unauthorized'}));
+    }
+    
+    // Route analytics requests
+    if (request.url.path.contains('analytics_content_audience')) {
+      final audienceType = request.url.queryParameters['audienceType'];
+      final result = await ParseAnalyticsEndpoints.handleAudienceRequest(audienceType ?? 'total_users');
+      return Response.ok(json.encode(result));
+    }
+    
+    if (request.url.path.contains('analytics')) {
+      final endpoint = request.url.queryParameters['endpoint'] ?? 'audience';
+      final from = int.tryParse(request.url.queryParameters['from'] ?? '0') ?? 0;
+      final to = int.tryParse(request.url.queryParameters['to'] ?? '0') ?? DateTime.now().millisecondsSinceEpoch ~/ 1000;
+      final stride = request.url.queryParameters['stride'] ?? 'day';
+      
+      final result = await ParseAnalyticsEndpoints.handleAnalyticsRequest(
+        endpoint: endpoint,
+        startDate: DateTime.fromMillisecondsSinceEpoch(from * 1000),
+        endDate: DateTime.fromMillisecondsSinceEpoch(to * 1000),
+        interval: stride,
+      );
+      return Response.ok(json.encode(result));
+    }
+    
+    if (request.url.path.contains('analytics_retention')) {
+      final at = int.tryParse(request.url.queryParameters['at'] ?? '0') ?? 0;
+      final cohortDate = at > 0 ? DateTime.fromMillisecondsSinceEpoch(at * 1000) : null;
+      
+      final result = await ParseAnalyticsEndpoints.handleRetentionRequest(cohortDate: cohortDate);
+      return Response.ok(json.encode(result));
+    }
+    
+    return Response.notFound('Endpoint not found');
+  };
+}
+''';
 }
