--- packages/flutter/lib/src/analytics/parse_analytics_endpoints_clean.dart
@@ -1,277 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'parse_analytics_clean.dart';
-
-/// HTTP endpoint handlers for Parse Dashboard Analytics integration
-class ParseAnalyticsEndpoints {
-  
-  /// Handle audience analytics requests for Parse Dashboard
-  static Future<Map<String, dynamic>> handleAudienceRequest(String audienceType) async {
-    try {
-      final userAnalytics = await ParseAnalytics.getUserAnalytics();
-      final installationAnalytics = await ParseAnalytics.getInstallationAnalytics();
-      
-      switch (audienceType) {
-        case 'total_users':
-          return {
-            'total': userAnalytics['total_users'],
-            'content': userAnalytics['total_users']
-          };
-          
-        case 'daily_users':
-          return {
-            'total': userAnalytics['daily_users'],
-            'content': userAnalytics['daily_users']
-          };
-          
-        case 'weekly_users':
-          return {
-            'total': userAnalytics['weekly_users'],
-            'content': userAnalytics['weekly_users']
-          };
-          
-        case 'monthly_users':
-          return {
-            'total': userAnalytics['monthly_users'],
-            'content': userAnalytics['monthly_users']
-          };
-          
-        case 'total_installations':
-          return {
-            'total': installationAnalytics['total_installations'],
-            'content': installationAnalytics['total_installations']
-          };
-          
-        case 'daily_installations':
-          return {
-            'total': installationAnalytics['daily_installations'],
-            'content': installationAnalytics['daily_installations']
-          };
-          
-        case 'weekly_installations':
-          return {
-            'total': installationAnalytics['weekly_installations'],
-            'content': installationAnalytics['weekly_installations']
-          };
-          
-        case 'monthly_installations':
-          return {
-            'total': installationAnalytics['monthly_installations'],
-            'content': installationAnalytics['monthly_installations']
-          };
-          
-        default:
-          return {'total': 0, 'content': 0};
-      }
-    } catch (e) {
-      if (kDebugMode) {
-        print('Error handling audience request: $e');
-      }
-      return {'total': 0, 'content': 0};
-    }
-  }
-
-  /// Handle analytics time series requests for Parse Dashboard
-  static Future<Map<String, dynamic>> handleAnalyticsRequest({
-    required String endpoint,
-    required DateTime startDate,
-    required DateTime endDate,
-    String interval = 'day',
-  }) async {
-    try {
-      String metric;
-      switch (endpoint) {
-        case 'audience':
-          metric = 'users';
-          break;
-        case 'installations':
-          metric = 'installations';
-          break;
-        default:
-          metric = endpoint;
-      }
-      
-      final requestedData = await ParseAnalytics.getTimeSeriesData(
-        metric: metric,
-        startDate: startDate,
-        endDate: endDate,
-        interval: interval,
-      );
-      
-      return {'requested_data': requestedData};
-    } catch (e) {
-      if (kDebugMode) {
-        print('Error handling analytics request: $e');
-      }
-      return {'requested_data': <List<num>>[]};
-    }
-  }
-
-  /// Handle user retention requests for Parse Dashboard
-  static Future<Map<String, double>> handleRetentionRequest({DateTime? cohortDate}) async {
-    try {
-      return await ParseAnalytics.getUserRetention(cohortDate: cohortDate);
-    } catch (e) {
-      if (kDebugMode) {
-        print('Error handling retention request: $e');
-      }
-      return {'day1': 0.0, 'day7': 0.0, 'day30': 0.0};
-    }
-  }
-
-  /// Handle billing storage requests for Parse Dashboard
-  static Map<String, dynamic> handleBillingStorageRequest() {
-    // Mock implementation - replace with actual storage calculation
-    return {
-      'total': 0.5, // 500MB in GB
-      'limit': 100,
-      'units': 'GB'
-    };
-  }
-
-  /// Handle billing database requests for Parse Dashboard  
-  static Map<String, dynamic> handleBillingDatabaseRequest() {
-    // Mock implementation - replace with actual database size calculation
-    return {
-      'total': 0.1, // 100MB in GB
-      'limit': 20,
-      'units': 'GB'
-    };
-  }
-
-  /// Handle billing data transfer requests for Parse Dashboard
-  static Map<String, dynamic> handleBillingDataTransferRequest() {
-    // Mock implementation - replace with actual data transfer calculation
-    return {
-      'total': 0.001, // 1GB in TB
-      'limit': 1,
-      'units': 'TB'
-    };
-  }
-
-  /// Handle slow queries requests for Parse Dashboard
-  static List<Map<String, dynamic>> handleSlowQueriesRequest({
-    String? className,
-    String? os,
-    String? version,
-    DateTime? from,
-    DateTime? to,
-  }) {
-    // Mock implementation - replace with actual slow query analysis
-    return [
-      {
-        'className': className ?? '_User',
-        'query': '{"username": {"regex": ".*"}}',
-        'duration': 1200,
-        'count': 5,
-        'timestamp': DateTime.now().toIso8601String()
-      }
-    ];
-  }
-}
-
-/// Express.js middleware generator for Parse Dashboard integration
-/// 
-/// Usage:
-/// ```javascript
-/// const express = require('express');
-/// const app = express();
-/// app.use('/apps/:appSlug/', getExpressMiddleware());
-/// ```
-String getExpressMiddleware() {
-  return '''
-// Parse Dashboard Analytics middleware
-function parseAnalyticsMiddleware(req, res, next) {
-  // Add authentication middleware here
-  const masterKey = req.headers['x-parse-master-key'];
-  if (!masterKey || masterKey !== process.env.PARSE_MASTER_KEY) {
-    return res.status(401).json({ error: 'Unauthorized' });
-  }
-  
-  // Route analytics requests
-  if (req.path.includes('analytics_content_audience')) {
-    // Handle audience requests
-    const audienceType = req.query.audienceType;
-    // Call your Flutter analytics endpoint here
-    return res.json({ total: 0, content: 0 });
-  }
-  
-  if (req.path.includes('analytics')) {
-    // Handle analytics time series requests
-    const { endpoint, stride, from, to } = req.query;
-    // Call your Flutter analytics endpoint here
-    return res.json({ requested_data: [] });
-  }
-  
-  if (req.path.includes('analytics_retention')) {
-    // Handle retention requests
-    // Call your Flutter analytics endpoint here
-    return res.json({ day1: 0, day7: 0, day30: 0 });
-  }
-  
-  next();
-}
-
-module.exports = parseAnalyticsMiddleware;
-''';
-}
-
-/// Dart Shelf handler for Parse Dashboard integration
-/// 
-/// Usage:
-/// ```dart
-/// import 'package:shelf/shelf.dart';
-/// import 'package:shelf/shelf_io.dart' as io;
-/// 
-/// void main() async {
-///   final handler = getDartShelfHandler();
-///   final server = await io.serve(handler, 'localhost', 3000);
-/// }
-/// ```
-String getDartShelfHandler() {
-  return '''
-import 'dart:convert';
-import 'package:shelf/shelf.dart';
-
-Response Function(Request) getDartShelfHandler() {
-  return (Request request) async {
-    // Add authentication here
-    final masterKey = request.headers['x-parse-master-key'];
-    if (masterKey != 'your_master_key') {
-      return Response.forbidden(json.encode({'error': 'Unauthorized'}));
-    }
-    
-    // Route analytics requests
-    if (request.url.path.contains('analytics_content_audience')) {
-      final audienceType = request.url.queryParameters['audienceType'];
-      final result = await ParseAnalyticsEndpoints.handleAudienceRequest(audienceType ?? 'total_users');
-      return Response.ok(json.encode(result));
-    }
-    
-    if (request.url.path.contains('analytics')) {
-      final endpoint = request.url.queryParameters['endpoint'] ?? 'audience';
-      final from = int.tryParse(request.url.queryParameters['from'] ?? '0') ?? 0;
-      final to = int.tryParse(request.url.queryParameters['to'] ?? '0') ?? DateTime.now().millisecondsSinceEpoch ~/ 1000;
-      final stride = request.url.queryParameters['stride'] ?? 'day';
-      
-      final result = await ParseAnalyticsEndpoints.handleAnalyticsRequest(
-        endpoint: endpoint,
-        startDate: DateTime.fromMillisecondsSinceEpoch(from * 1000),
-        endDate: DateTime.fromMillisecondsSinceEpoch(to * 1000),
-        interval: stride,
-      );
-      return Response.ok(json.encode(result));
-    }
-    
-    if (request.url.path.contains('analytics_retention')) {
-      final at = int.tryParse(request.url.queryParameters['at'] ?? '0') ?? 0;
-      final cohortDate = at > 0 ? DateTime.fromMillisecondsSinceEpoch(at * 1000) : null;
-      
-      final result = await ParseAnalyticsEndpoints.handleRetentionRequest(cohortDate: cohortDate);
-      return Response.ok(json.encode(result));
-    }
-    
-    return Response.notFound('Endpoint not found');
-  };
-}
-''';
-}
